test: update perf tests to use esm
