+++ packages/mobx/scripts/perf.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-time node --expose-gc ./__tests__/perf/index.js $1
+mv ./dist/mobx.esm.production.min.js ./dist/mobx.esm.production.min.mjs
+
+time node --expose-gc ./__tests__/perf/index.mjs $1
+
+mv ./dist/mobx.esm.production.min.mjs ./dist/mobx.esm.production.min.js
